refactor: betting 패키지로 BettingRate와 BettingResult 이동
