+++ src/main/java/blackjack/model/betting/BettingRate.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.betting;
 
 import blackjack.model.state.State;
