+++ src/main/java/blackjack/model/betting/BettingResult.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.betting;
 
 import blackjack.model.participant.Participant;
 import java.util.LinkedHashMap;
